multiagent-fin/backend/stt_service/requirements.txt
fastapi
uvicorn
openai-whisper